Fishing fixes thanks to Carlos_Ed. Quote: I made several fixes to the fishing procedure in L2J.
The current code in 1283 SVN revision has a few bugs, and some checks are sort of miss-placed in my opinion. These made fishing work flawlessly, and has been tested extensively. Also, I made a few adds to be able to extract zone coordinates so that I could improve the hook landing position.

Before, the hook had a fixed height right about the same level that the PC's foot. Now it extracts the water surface Z from the zone info, and places it on the water surface. The check to see if some coordinates were inside a fishing zone were also changed to allow casting the rod from bridges and other higher places. You can now fish from high rocks, balconies, peers and bridges. The hook will be cast if the water zone is in range and it will land in water, and you will see the string going from the fishing rod, down. I also added a check to see if the PC was not inside the fishing zone itself, thus allowing to have fishing zones a bit higher that "sea level" and still prevent "underwater fishing" on such zones.

Some code was moved off L2PcInstance. Namely the zone checks, and the hook landing coordinates selection, from startFishing(). The old method caused some problems, the most noticeable was that whenever you used the skill one lure was always consumed, no matter if fishing actually took place or not.


git-svn-id: http://svn.l2jserver.com/trunk@1285 d34dfa0f-f91c-0410-86dd-cf06f19dd0ad