Little fix for //quest_reload. NOTE, This commit does NOT fix the following issues: a) the quests may still ONLY be reloaded if the questname starts with an alpha character (i.e. non-number) b) npc reloads (respawn-all etc) still cause all NPCs to lose all their quests c) there is still an issue when removing an NPC from a quest that you reload; namely, if the new version of the quest is removing an NPC from an event (like deleting a QUEST.addTalkId(xxxxx) line), the npc will NOT be removed!  That particular NPC will, in fact, still use the old version of the quest!!  (this will be fixed in a future update)
This commit DOES fix the following:
a) The players' QuestStates no longer refer to the old scripts.  Instead, the new quest scripts are always found correctly from the QuestState
b) If a mobs that are registered by the new version, including both mobs that also existed in the old version and mobs that were newly adde to the script, will properly update their quest references and work as dictated by the new script.
c) The QuestManager class properly keeps track only of the latest quest, using its questName as lookup key.

This has been tested by Emperorc and by myself and worked to the above specs so far.  More to come soon...


git-svn-id: http://svn.l2jserver.com/trunk@1427 d34dfa0f-f91c-0410-86dd-cf06f19dd0ad