Cursed Weapons:  DP side for [DP3872] (thanks Ahmed, Baorc & evil33t)
Bug Corrected:
When cursed weapon is equipped by a player and then is dropped on the ground, the cursed weapon does not lose all of the previous levels that the previous owner had done.
- This has been corrected so it does not retain the levels set by the previous owner.
Kamael Cursed Weapon Transformation:
When a player equips the cursed weapons they transform into the associated transformation. A players name will display to the player as his own name, but to others it will display the name associated with the cursed weapon. In the community board the player will display as the players true name, also in ALL chat the player will display the name of the associated cursed weapon, party is not allowed, shout and trade are disabled for cursed weapon users, and clan chats should be displaying the true characters name.

git-svn-id: http://svn.l2jserver.com/trunk@1521 d34dfa0f-f91c-0410-86dd-cf06f19dd0ad