Large patch: Symbol/signet skills and char heading calculation fixes.
Symbol skills big thanks to Nuradi for much of the code and Rnn for info and testing.
Added or reworked the following skills:
Volcano, Cyclone, Gehenna, RagingWaves, Anti-Summoning-Field, Symbol of the Assassin
Symbol of the Sniper, Symbol of Energy, Symbol of Honor, Symbol of Resistance, 
Symbol of Noise, Symbol of Defense, Battle Force, Spell Force

Some DP corrections remain.

Char heading calculation came up when testing this patch and I ended up testing and 
fixing isBehind, changing "isFrontTarget" to isInFrontOf(target) and isFacing(target). 
Also char heading was not changed server side when attack or cast started so no wonder 
there was errors in this.

git-svn-id: http://svn.l2jserver.com/trunk@1836 d34dfa0f-f91c-0410-86dd-cf06f19dd0ad