Alternative DataTables...
This was made by Vice for L2JOneo,and now he's worked his magic and brought it to you on L2J....thanks a lot Vice, been using this ever since you made it!

Before I explain all this, I want to stress the fact that non of these tables will have any effect on your server if you have them disabled from the configurations, so that means nothing is loaded from them, and you can have that table deleted and not in your database and it won't have any negative effect.

Basically almost all server admins do some kind of changes to their L2J database, and when they want to update their database they always have to worry about conflicting ID's, and issues where they could lose some data they made, i.e spawns, npcs, etc.

Well this issue doesn't exist with this commit because when you put your custom data inside these tables you won't ever have to worry about the L2JDP team making a new update to them. These files are to be completely empty and never touched by L2JDP team unless if it is to add/remove appropriate columns, or to change things such as truncation. 
Yes, this meaning L2JDP doesn't add any data or information here, and just modifies the structure in appropriate situations.

Explaining the tables:

custom_armor -> If you modified the stats of an armor you can put the armor here and delete it from the original armor table and safely have it loaded and never touched by l2j.

custom_armorsets -> same as above description.

custom_droplist -> same as above description.

custom_etcitem -> same as above description.

custom_merchant_buylists -> same as above description.

custom_merchant_shopids -> same as above description.

custom_notspawned -> When DeleteGmSpawnOnCustom is true, this will make it so when a GM deletes an NPC in game, it does not actually get deleted from the spawnlist, it gets its ID put into a list of ID's that are NOT loaded with the rest of the spawns, this way you can prevent loss of spawn data by simply making it "not load that id", to load it again, simply go in this table and remove the id from the list. Id's with "isCustom = 0", are spawns from the spawnlist table, and Id's with "isCustom = 1", are spawns from the custom_spawnlist table.

custom_npc -> If you added a custom npc, or modified the stats of existing npc you can put the npc here and delete it from the original npc table and safely have it loaded and never touched by l2j. In cases where you created your own npc and Id, you simply don't have to worry about deleting the ID unless if there is an ID in the NPC table that is using that ID.

custom_spawnlist -> here it is the same as above, but you do not need to worry about having the same id in the spawnlist table and the custom_spawnlist table

custom_teleport -> same as custom_npc

custom_weapon -> same as custom_armor

(Yeah I plagerized all this from my original thread on the forum.... so if you've read it twice now I can but only laugh if you didn't realize that and commend you if you did and continued to read my beautiful explanations :))

git-svn-id: http://svn.l2jserver.com/trunk@1910 d34dfa0f-f91c-0410-86dd-cf06f19dd0ad