Thanks a lot House,
Quoting House:

Current implementation of FlyToLocation effects don't allow to configure some important variables that define the flight.

For example, you will always flight backwards your target in case you use skill Blink but you should flight your own backwards or in case you're trying to use Warp (not implemented yet due to these problems) . So here is a patch that temporary fix these limitations.

First of all change effect's name to Warp instead of Blink, because Blink seems to be a combination of some effects, including warp. So EffectBlink.java is deleted. A new file EffectWarp, based on EffectBlink.java is added and also a new effect type: WARP.

Secondly: there is one more stat for skills: int flyCourse. It means flying course and works this way: Imagine a compass aboue your character in which your heading is set as North. If you wanna fly backwards, fly course will have to be 180. If you wanna fly towards, flycourse will have to be 360 or 0.

Finally: now effect will recognized who is the flyer, target or caster. This is important, for example, if NcSoft decides to create a new skill that makes target flying around xD. Now, you will have to choose is EffectWarp is self effect or not.

This is an example, skill Blink:

{{{
<skill id="1448" levels="1" name="Blink">
<!--
  Description:
  Teleports you backwards while stunning and damaging nearby enemies. Consumes 3 Souls.
 
-->
  <set name="mpInitialConsume" val="30"/>
  <set name="mpConsume" val="70"/>
  <set name="power" val="1000"/>
  <set name="hitTime" val="300"/>
  <set name="operateType" val="OP_ACTIVE"/>
  <set name="isMagic" val="true"/>
  <set name="target" val="TARGET_AURA"/>
  <set name="skillRadius" val="150"/>
  <set name="skillType" val="PDAM"/>
  <set name="castRange" val="-1"/>
  <set name="effectRange" val="-1"/>
  <set name="reuseDelay" val="36000"/>
  <set name="soulConsumeCount" val="3"/>
  <set name="effectPower" val="50"/>
  <set name="effectType" val="STUN"/>
  <set name="flyRadius" val="300"/>
  <set name="flyCourse" val="180"/>
  <for>
    <effect count="1" name="Stun" time="9" val="0" stackOrder="1" stackType="Stun"/>
    <effect self="1" count="1" name="Warp" time="1" val="0"/>
  </for>
</skill>
}}}


git-svn-id: http://svn.l2jserver.com/trunk@2071 d34dfa0f-f91c-0410-86dd-cf06f19dd0ad